third_party/WebKit/LayoutTests/presentation/presentationrequest.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/gc.js"></script>
<script>

const presentationUrl = 'http://example.com';
const presentationUrls = [presentationUrl, 'cast://google.com/app_id=deadbeef'];

const expectedException = new DOMException(
    'PresentationRequest::start() requires user gesture.',
    'InvalidAccessError');

promise_test(function(t) {
  const request = new PresentationRequest(presentationUrl);
  return promise_rejects(t, expectedException, request.start());
}, 'Test that the PresentationRequest.start() with one URL requires user gesture.')

promise_test(
    function(t) {
      const request = new PresentationRequest(presentationUrls);
      return promise_rejects(t, expectedException, request.start());
    },
    'Test that the PresentationRequest.start() with multiple URLs requires user gesture.')

    const testGarbageCollection = function(requestArgument) {
      navigator.presentation.defaultRequest =
          new PresentationRequest(requestArgument);
      navigator.presentation.defaultRequest.onconnectionavailable = function() {
      };
      gc();
      assert_not_equals(
          navigator.presentation.defaultRequest.onconnectionavailable,
          undefined);
    };

test(function() {
  testGarbageCollection(presentationUrl);
}, 'Test that navigator.presentation.defaultRequest.onconnectionavailable with one URL isn\'t reset after gc().');

test(function() {
  testGarbageCollection(presentationUrls);
}, 'Test that navigator.presentation.defaultRequest.onconnectionavailable with multiple URLs isn\'t reset after gc().');

test(function() {
  const request = new PresentationRequest('http://example.com');
  const promise_1 = request.getAvailability();
  const promise_2 = request.getAvailability();
  assert_true(promise_1 === promise_2);
}, 'Test that the PresentationRequest.getAvailability() returns same promise object.');

</script>
</body>
</html>
